Use arg as data dir for merge_binpacks, fix filter plain

+++ merge_binpacks.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 # Merges all .binpack files in a directory
+if [ "$#" -ne 1 ]; then
+  echo "Usage: ./merge_binpacks.sh <data_dir>"
+  exit 0
+fi
 
-data_dir=/dev/shm
 output_binpack=merged.binpack
 
 rm -rf $output_binpack
 python3 ./stockfish/script/interleave_binpacks.py \
-  $(ls -1 $data_dir/*.plain.binpack | tr '\n' ' ' && echo $output_binpack)
+  $(ls -1 $1/*.plain.binpack | tr '\n' ' ' && echo $output_binpack)
 ls -lth $output_binpack
